optimize pwa
