+++ src/app/layout.tsx
@@ -1,4 +1,5 @@
 "use client";
+import "./globals.scss";
 import { MetaMaskUIProvider } from "@metamask/sdk-react-ui";
 import NextAuthSessionProvider from "./providers/sessionProvider";
 import { Session } from "next-auth";
